+++ include/kernel/domain/alloc/page_alloc.h
@@ -5,18 +5,18 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- * 
+ *
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- * 
+ *
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,15 +29,25 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef JINUE_KERNEL_SERVICES_VMALLOC_H
-#define JINUE_KERNEL_SERVICES_VMALLOC_H
+#ifndef JINUE_KERNEL_DOMAIN_PAGE_ALLOC_H
+#define JINUE_KERNEL_DOMAIN_PAGE_ALLOC_H
 
 #include <kernel/types.h>
 
-addr_t vmalloc(void);
+#define PFNULL ((kern_paddr_t)-1)
 
-void vmfree(addr_t page);
+void *page_alloc(void);
 
-bool vmalloc_is_in_range(addr_t page);
+void page_free(void *page);
+
+unsigned int get_page_count(void);
+
+bool add_page_frame(kern_paddr_t paddr);
+
+kern_paddr_t remove_page_frame(void);
+
+void clear_page(void *page);
+
+void clear_pages(void *first_page, int num_pages);
 
 #endif
