+++ kernel/domain/alloc/page_alloc.c
@@ -31,8 +31,8 @@
 
 #include <jinue/shared/asm/machine.h>
 #include <jinue/shared/asm/syscall.h>
-#include <kernel/domain/services/page_alloc.h>
-#include <kernel/domain/services/vmalloc.h>
+#include <kernel/domain/alloc/page_alloc.h>
+#include <kernel/domain/alloc/vmalloc.h>
 #include <kernel/machine/vm.h>
 #include <string.h>
