+++ kernel/domain/alloc/slab.c
@@ -29,9 +29,9 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+#include <kernel/domain/alloc/page_alloc.h>
+#include <kernel/domain/alloc/slab.h>
 #include <kernel/domain/services/logging.h>
-#include <kernel/domain/services/page_alloc.h>
-#include <kernel/domain/services/slab.h>
 #include <kernel/machine/cpu.h>
 #include <kernel/utils/utils.h>
 #include <kernel/types.h>
