+++ kernel/domain/alloc/vmalloc.c
@@ -5,18 +5,18 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
  * are met:
- *
+ * 
  * 1. Redistributions of source code must retain the above copyright
  *    notice, this list of conditions and the following disclaimer.
- *
+ * 
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in the
  *    documentation and/or other materials provided with the distribution.
- *
+ * 
  * 3. Neither the name of the author nor the names of other contributors
  *    may be used to endorse or promote products derived from this software
  *    without specific prior written permission.
- *
+ * 
  * THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS "AS IS" AND
  * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
  * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -29,25 +29,56 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef JINUE_KERNEL_SERVICES_PAGE_ALLOC_H
-#define JINUE_KERNEL_SERVICES_PAGE_ALLOC_H
+#include <jinue/shared/vm.h>
+#include <kernel/domain/alloc/vmalloc.h>
+#include <kernel/domain/services/panic.h>
 
-#include <kernel/types.h>
 
-#define PFNULL ((kern_paddr_t)-1)
+/**
+ * @file
+ *
+ * Virtual address allocator
+ *
+ * This allocator (specifically vmalloc()) is used to allocate a free virtual
+ * address in the kernel's address space to which a page frame can be mapped.
+ *
+ * If you want to allocate a mapped page ready to use, use the page allocator
+ * instead, i.e. page_alloc().
+ *
+ * */
 
-void *page_alloc(void);
+/**
+ * Allocate a page of virtual address space.
+ *
+ * @return address of allocated page or NULL if allocation failed
+ *
+ * */
+addr_t vmalloc(void) {
+    /* TODO implement this */
+    panic("vmalloc(): not implemented");
 
-void page_free(void *page);
+    /* never reached */
+    return NULL;
+}
 
-unsigned int get_page_count(void);
+/**
+ * Free a page of virtual address space.
+ *
+ * @param page the address of the page to free
+ *
+ * */
+void vmfree(addr_t page) {
+    /* TODO implement this */
+    panic("vmfree(): not implemented");
+}
 
-bool add_page_frame(kern_paddr_t paddr);
-
-kern_paddr_t remove_page_frame(void);
-
-void clear_page(void *page);
-
-void clear_pages(void *first_page, int num_pages);
-
-#endif
+/**
+ * Check whether the specified page is in the region managed by the allocator.
+ *
+ * @param page the address of the page
+ * @return true if it is in the region, false otherwise
+ *
+ * */
+bool vmalloc_is_in_range(addr_t page) {
+    return is_kernel_pointer(page);
+}
