+++ kernel/infrastructure/elf.c
@@ -32,14 +32,14 @@
 #include <jinue/shared/asm/machine.h>
 #include <jinue/shared/asm/syscall.h>
 #include <jinue/shared/vm.h>
+#include <kernel/domain/alloc/page_alloc.h>
+#include <kernel/domain/alloc/vmalloc.h>
 #include <kernel/domain/entities/descriptor.h>
 #include <kernel/domain/entities/object.h>
 #include <kernel/domain/entities/process.h>
 #include <kernel/domain/services/cmdline.h>
 #include <kernel/domain/services/logging.h>
-#include <kernel/domain/services/page_alloc.h>
 #include <kernel/domain/services/panic.h>
-#include <kernel/domain/services/vmalloc.h>
 #include <kernel/infrastructure/elf.h>
 #include <kernel/machine/auxv.h>
 #include <kernel/machine/vm.h>
