+++ kernel/infrastructure/i686/thread.c
@@ -29,7 +29,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <kernel/domain/services/page_alloc.h>
+#include <kernel/domain/alloc/page_alloc.h>
 #include <kernel/infrastructure/i686/asm/thread.h>
 #include <kernel/infrastructure/i686/cpu.h>
 #include <kernel/infrastructure/i686/cpu_data.h>
