+++ kernel/infrastructure/i686/vm_x86.c
@@ -29,7 +29,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <kernel/domain/services/vmalloc.h>
+#include <kernel/domain/alloc/vmalloc.h>
 #include <kernel/infrastructure/i686/boot_alloc.h>
 #include <kernel/infrastructure/i686/vm_private.h>
 #include <assert.h>
